server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7